Clean up and improve build scripts (#46)
* Clean up and improve build scripts

The scripts now install all packages in one command. This removes the
repeated execution of the same command.

If CMake is not installed, the system CMake is used. This allows the
scripts to be used on VMs without a pre-built CMake.

* Clean up and improve build scripts

The scripts now install all packages in one command. This removes the
repeated execution of the same command.

If CMake is not installed, the system CMake is used. This allows the
scripts to be used on VMs without a pre-built CMake.

* remove systemtap-sdt-devel rpm-build from zypper install

* remove 'set' which overrides exit code

* restore zypper install of rpm-build