Fix Galera configuration files
The writeset limits can cause tests to fail if the writeset size or
rowcount exceeds the configured limits. For this reason, it is better to
have no row limits and use the default writeset limit of 2GB.

Adding wsrep_causal_reads also removes some of the uncertainty of reads
with Galera clusters by enforcing stricter synchronization semantics.